feat: LocalPath also sets group (if not specified) based on user (#38)
This PR aligns the behaviour of the file creation methods (mkdir, write_bytes,
write_text) when the user argument is provided but the group argument is not.

In this case, Pebble sets the group based on the user (like chown $user:
$path), while shutil.chown by default does not (like chown $user $path). Since
changing the Pebble behaviour to match shutil here would require additional API
calls (and possibly creating the directory twice?), and since the user and
group arguments for the file creation methods are an API design that follows
Pebble rather than pathlib, LocalPath is updated to follow Pebble's behaviour
here.

This PR updates the unit tests to reflect this change, and adds a collection of
integration tests for file ownership.

Documentation of the error case when group is provided without user for
ContainerPath has also been added.

Resolves #39