fix: write methods no longer change existing mode unless it's passed explicitly (#47)
The current behaviour of both `ContainerPath` and `LocalPath` for
`write_{bytes,text}` is to always set the mode -- even setting the default mode
if none is provided. I think overriding any existing file permissions with the
default when none is provided is a mistake, as (e.g.)
`pathlib.Path.write_bytes(data)` doesn't change any existing file permissions.
This PR modifies these methods so that the permissions are only changed if the
`mode` argument is not `None`.

Tests have been expanded to cover the case of writing when the target file
already exists, and updated to reflect this new behaviour.

Resolves #46