fix: consistent behaviour for mkdir chown args when exists (#48)
This PR makes the behaviour of the `user` and `group` args to the file creation
methods consistent.

- If the file already exists, `ContainerPath.write_{bytes,text}` will no longer
  modify the `user` or `group` if they're not supplied explicitly, just like
  how we recently updated these methods to not change the `mode` if it's not
  supplied explicitly in #47 
- If the directory already exists, `LocalPath.mkdir` will no longer modify the
  `user` or `group`, even if they're passed explicitly, matching the current
  `ContainerPath.mkdir` behaviour, and matching the behaviour of `mode` for
  `mkdir`.

This PR also adds Juju integration tests to cover the `user` and `group`
behaviour when the target file or directory already exists, and updates the
unit tests to work with the new behaviour.