ci: multi-base packing (#56)
This PR adds the option to specify the base when packing. Currently the 20.04
and 24.04 bases are supported and exercised in CI. The charm + pathops code is
therefore now exercised on both bases in the integration tests run in CI. Most
significantly this means that pathops is tested with multiple Python versions
in integration tests (3.8 and 3.12).