docs: simplify docs makefile and use just for docs (#75)
This PR refactors the docs CI to use uv to run the sphinx commands, and use the
same command to build the docs in readthedocs as is used locally.

I've also taken the opportunity to move the commands to a justfile for cleaner
syntax and the ability to run the docs build commands from anywhere in the
repository. They're all in a separate module, docs.just, keeping things clean
and separate. This allows the default docs build (html, used in readthedocs) to
be run with just docs, while other commands can be run with (e.g.) just docs
linkcheck.

The main _docs/Makefile targets are preserved for compatibility with the
Canonical docs workflows, and to allow contributors who are used to this format
to easily find and run the docs commands. These targets run just, making the
docs.just file the single source of truth in all cases.