docs: add remaining general libraries to listing (#79)
This PR adds the other known general libraries to the library listing docs. The
libraries were found using the methodology described in #73, where only a small
number of the known libraries were added to simplify review.

The emojis for the statuses are updated based on discussion, as are status
tooltips and library sorting. Key tables for the emojis now only include the
emojis used on that page. Libraries with the `'unlisted'` status are filtered
out of the table. Charmcraft-hosted libs are prefixed with their charm name,
and linebreaks are controlled with injected html. 

A small number of words are added to the custom wordlist.