ci: unify pebble and ubuntu test suites as functional (#115)
This PR combines the non-Juju integration tests (`integration/pebble` and `integration/ubuntu`) into the `functional` test suite (alongside `unit` and `integration`). Juju integration tests are moved from `integration/juju` to `integration`.

To facilitate the different needs of different packages, the `functional` test suite can be customised in CI -- the package can specify the matrix of `pebble` versions to use (or not use it at all), the matrix of Ubuntu versions to run on, and whether or not the tests need to be run with `sudo`. The default is to use `ubuntu-latest` without `sudo`, and not to install `pebble`.