ci: release on version change (#118)
This PR changes our publication workflow from running on the push event of specific tags (one per library), to running on every push to main. The unify-publish-inputs.py script has been modified to:

1. Get a list of the packages that changed in this event
2. Narrow that list to just those that have a different version now (new packages are included as having a different version -- the first merge to main will be a release, so they should start with a 0.X + .devY / .aY version unless they really are ready for a true release)

The publish job is then run with just those packages (using the matrix strategy).

Narrowing the publish job to only run on packages with a changed version is mostly an efficiency and UX thing -- running the publish job with a package whose version is identical to a released version will fail for that package, but there shouldn't be any other bad effects.

This release strategy has the following benefits:

- We don't need to worry about permissions to use github releases or push tags to main
- Adding a new library won't require editing the publish job to add its tag
- CODEOWNERS for a library control its releases via their approval and merge
- We don't need to worry about limitations of the push tag event, like how it won't trigger at all if more than 3 tags are pushed at once