ci: add CODEOWNERS file (#133)
This PR adds the `CODEOWNERS` file that will be used to allow library
owners to approve PRs that affect just their libraries.
`@canonical/charm-tech` is set as the default owner for the repository
(`.`), as well as set explicitly for the `CODEOWNERS` file itself, the
repository infra files (e.g. `/pyproject.toml`, `/justfile`), global
docs, and the packages that we currently own. PRs adding a new library
will add a `CODEOWNERS` entry for its package directory, which will
require `@canonical/charm-tech` approval. After that, the owners of the
package directory will be able to approve changes on their own.