ci: add a static job to check if tests matrix passed/skipped (#136)
This PR adds a static `tests-passed` job to the `ci` workflow to check
if the dynamically generated jobs from the tests matrix passed (or were
skipped). This is necessary because the branch protection rules require
a known job ID to be specified when setting required checks (which block
merging PRs). Not only do we dynamically run different jobs depending on
what tests a package has (etc), but we may skip running them altogether
if a package hasn't been changed in that PR.