ci: switch from charm-tech to charmlibs-maintainers (#137)
This PR switches ownership of everything currently specified in the
`CODEOWNERS` file from `charm-tech` to `charmlibs-maintainers`. The
`charmlibs-maintainers` team is currently the exact same people, but the
code review settings are different: only I will be requested for
reviews, so the rest of the team should not get spammed. However, the
rest of the team still has code owner permissions, and can provide the
required approvals to merge PRs affecting the files the team owns.