ci: don't include removed packages in the list of packages to release (#143)
We release packages when a merge to main contains a version change for
that package, excluding `.dev` releases. This includes the case when a
new package is added (old version is `None`). However, it also
erroneously includes the case where a package is *removed* from the
repository (*new* version is `None`). This would fail anyway when we try
to run the release script, since the package files are not longer
present. For sanity, this PR excludes removed packages from the list of
packages to release.