ci: simplify (and correct?) handling of dependency groups (#154)
It's painful trying to use dependency groups from both the repo-level
`pyproject.toml` and the individual package `pyproject.toml`. We
currently work around this by dropping down to the `uv venv` and `uv
pip` commands instead of using `uv run`, which is a bit nasty, and may
not actually be working properly.

If we instead specify the tool versions in the `justfile`, things are a
lot simpler -- we can `uv run` using `--with` for our tool versions and
`--group` for the package's dependency group(s). This should also make
it clearer to library contributors where the `pyright` (etc) versions
are coming from.

This requires packages to specify all dependency groups, empty if they
don't need to put anything in them (which is how the template will start
new projects). This PR updates the existing packages to do this.