docs: correctly build package reference docs with their own deps (#152)
The docs build process is currently broken, because we try to install
deps for one package at a time and build one-by-one, but we leave the
automodule files in place as we do this incrementally, so once we add
another package with dependencies this breaks. This was noticed in #128.

This PR modifies the build process to first build the package reference
docs truly one-by-one. Rather than relying on the package reference docs
being left in place because of the automodule files sticking around, we
explicitly save the doctrees and their in-memory domain-specific
indexing (for cross-refs) to a file. These are then loaded when sphinx
tries to load these the package reference docs on subsequent iterations.

The `docs.just` just module has been cleaned up accordingly.