ci: accept a customisable list of tags to run integration tests with (#141)
The current integration test machinery is still a bit tied to the needs
of `charmlibs-pathops`, particularly in the hard-coded selection of
Ubuntu bases the integration test matrix uses. The base handling is also
a little bit funny in that it doesn't actually determine anything in
particular for CI -- we run on `ubuntu-latest` and Juju takes care of
ensuring charms runs on their requested base when deployed. Instead we
just pass the `base` value to the packing job and integration test job
via an environment variable (`pathops` uses it for packing and ignores
it for integration tests).

This PR changes this from a `base` argument to a `tag` argument, and
allows packages to define their own custom list of `tags` in their
`tool.charmlibs.integration` table. In CI, their integration tests are
run with a matrix of these `tags` and their supported substrates
(inferred from the `pytest` markers in their tests). The default
behaviour is to run with a single empty tag ([example
run](https://github.com/canonical/charmlibs/actions/runs/17375605646/job/49321143981?pr=141)
with `pathops` doing this). Packages can use this tag to specify the
base they want to pack their charm for (as `pathops` does), or for other
information, for example selecting different sets of tests (via
`skipif`), deploying with different charms, etc.

The name `tag` for this could perhaps be better. `label` clashes with
the `pytest` concept. `id` might be a good alternative.