docs: add a package template, an example library, and a job to keep them in sync (#144)
This PR adds a `cookiecutter` package template, under `.template/`.
Files for a new charmlib can be created with `just init`. Required
arguments will be prompted for, or they can be passed on the commandline
for faster testing/etc (see cookiecutter CLI help). A rendered version
of the template is also added, under `.example/`. The example is covered
by linting and tests (on changes, like all other packages). A CI job
ensures that the example is always kept in sync with the template.