fix: allow integer revision in Snap.ensure
Mostly a type checking fix with some subtleties -- if revision=0 was
passed, this would be treated as revision='' before, now will be
treated as revision='0' (unlikely to come up in the wild). Previously,
if the snap was installed with revision 123, revision=123 would compare
inequal to the stored revision string (123 != '123') and refresh would
be called unnecessarily.