feat: consistently allow revision to be int or str in public methods (#165)
`snap.ensure` requires `revision` to be `int | None`, while `snap.add`
and `snap.Snap.ensure` require it to be `str | None`. We consistently
use `str` internally (with `''` as the default value rather than
`None`). This PR allows the three public methods to accept `str | int |
None` for better consistency with each other -- it's a backwards
compatible change because it only widens the accepted type.

There's now a small behaviour change if the integer `0` is used for
`snap.add` or `snap.Snap.ensure` -- previously this would be a type
checking error, but would be converted to `''` internally at runtime.
Now it will become the string `'0'`. This seems more correct than
converting it to `''`. I couldn't find any charms using `revision=0`.