ci: replace functional-pebble recipe with setup and teardown scripts (#172)
How we handle running `functional` tests with `pebble` has always been a
bit of a CI wart, and this became very evident when working on the
tutorial in #159. This PR replaces the separate `functional` and
`functional-pebble` recipes with sourcing `tests/functional/setup.sh`
and `tests/functional/teardown.sh` in `just functional` before and after
running tests.

The primary motivation for this change is to have a simpler and more
consistent CI and task runner interface for most developers. This PR
does not add these files to the template, as most libraries won't need
them -- we'll cover this functionality in the docs soon (not the
tutorial), and just point to `pathops` as the example.