docs: add tutorial (#159)
This PR adds a tutorial: [preview
build](https://canonical-ubuntu-documentation-library--159.com.readthedocs.build/charmlibs/tutorial/).
It also adds a checked-in copy of the tutorial code, which is the
template code with some additions, and updates the list of directories
counted as packages to include `.tutorial` so it's tested in CI if it
changes.

The tutorial links to the old how-to to explain `git` hosted Python
packages, so I've also added a note to the how-to temporarily to
document that it's outdated. Updating the how-to properly will be a
separate PR.

---------

Co-authored-by: Dave Wilding <tech@dpw.me>