ci: unify library discovery (#177)
This PR replaces the current scattered and ad-hoc implementations of
package discovery in the monorepo for testing, docs, and publishing with
a single script that they all invoke.

## Motivation

Currently, the rules for iterating over package directories are
implemented in 4 different places:
1. CI
2. publishing
3. docs CI
4. a docs extension 

This isn't entirely problematic, as there are some slight differences in
the different use cases:
- collecting changed packages to test
- collecting bumped packages to publish
- collecting the packages that should be documented

However, it's now at the point where the current scattered
implementations would each require updates to facilitate adding
interface specifications without their libraries (#174), and the logic
would need to be re-implementation in more places for other planned
features (interface docs, changelog updates).

## Changes

This PR moves all the package discovery logic to a single script, with
options for whether to output all packages, only changed packages, or
only packages with version bumps. These were previously implemented in
two separate scripts with significant overlap. It's good to have them in
the same place as they each build incrementally on the previous.

There are also flags for filtering out the namespace placeholder
packages (used by docs), and filtering out the example packages (used by
docs and publishing), which was previously done in an ad-hoc basis in
several places.

The unified script outputs the files as a `json` list, to make it easy
to consume in other Python scripts (currently), and directly in
workflows (planned).

## Testing

I've verified these changes by making [a PR in my
fork](https://github.com/james-garner-canonical/charmlibs/pull/7), and
by [pushing the changes to my fork's
main](https://github.com/james-garner-canonical/charmlibs/actions/runs/18000916594),
as well as by executing the new script locally, and manually running
`pyright` on it (our scripts are not currently covered by automatic type
checking).

I wonder if it would be worth adding tests for some of these scripts to
our CI.

---------

Co-authored-by: Tony Meyer <tony.meyer@gmail.com>