feat: add tracing interface spec to charmlibs monorepo (#185)
This PR copies the [tracing interface definition from
charm-relation-interfaces](https://github.com/canonical/charm-relation-interfaces/tree/main/interfaces/tracing)
to this repo, under the `interfaces/tracing` directory. Some small
automatic changes have been made, but have been kept to a minimum with
the addition of a `ruff.toml` file to configure linting rules. One long
comment line was manually broken.

This PR will be merged without review so that we have a starting point
of the interface definition as-is. Docs and tests can be updated as we
add machinery to exercise them.