ci: fix lint recipe variadic argument passing (#198)
This PR fixes a bug in the interaction between the `lint` and `static`
recipes, and improves the output of the `lint` recipe.

The `lint` and `static` recipes accept variadic arguments to pass to
`pyright`. Currently, if we run `just lint $package` (with no variadiac
args), the `static` is called as `just static $package ''`, resulting in
a call like `pyright ''` -- that is, a single empty string as the only
positional argument. This results in `pyright` type checking the current
directory (desired) but ignoring the local config.

This PR fixes this issue like we've done with other cases where variadic
arguments need to be passed to different recipes, by switching from
using `just` dependencies to passing the arguments positionally to the
script.

Additionally, I've taken the opportunity to improve the output of
`lint`, so that a linting failure in `ruff` or `codespell` doesn't
result in `pyright` being skipped.