ci: run interface tests (#195)
This PR adds the ability to run interface tests to our CI.

## Context

The `charmlibs` monorepo is the new home for interface definitions and
tests. We need to be able to execute existing interface tests in our CI.

Interface test execution in `charm-relation-interfaces` is opaque and
very abstract. I've attempted to improve this in a couple of PRs (see
below), which have informed my approach here:
- https://github.com/canonical/charm-relation-interfaces/pull/292)
- https://github.com/canonical/charm-relation-interfaces/pull/297

The approach is:
1. Separate test setup, execution, and teardown from charm and test
discovery.
2. Run interface tests with `pytest` as directly as possible.
3. Execute each `pytest` invocation in a separate job (one per
`interface / version / role / charm / endpoint` combination).

## Overview

This PR adds a new `interface-tests` workflow.

- It runs interface tests for a single interface, totally compatible
with existing interface tests and charm specific interface test
configuration.
- It's called from the `ci` workflow following the same approach as we
use for the `tests` workflow.
- It runs a separate matrix job for each `(version / role (provide or
require) / charm / endpoint)` combination known for the interface.

Two scripts have been added to support this workflow.
- `.scripts/get-interface-test-targets.py` calculates the `(version /
role / charm / endpoint)` targets for the matrix
- `.scripts/run-interface-tests.py` runs the interface tests for a
single target

These scripts replace
[run_matrix.py](https://github.com/canonical/charm-relation-interfaces/blob/main/run_matrix.py)
from `charm-relation-interfaces` and the interface test and endpoint
discovery and execution parts of
[pytest-interface-tester](https://github.com/canonical/pytest-interface-tester/tree/main/interface_tester).

## Implementation details

Interface tests are discovered in the local `charmlibs` repository,
written to a cloned copy of the charm repo and then executed with
`pytest`, using an `interface_tester` based fixture used to configure,
setup, and teardown the interface tests (which are currently written
using a pseudo-singleton
[Tester](https://github.com/canonical/pytest-interface-tester/blob/main/interface_tester/interface_test.py#L175)
class). In combination with execution in a separate Github job, this
provides much better output in CI.

Currently, the `interface_tester` library only provides the
`InterfaceTester.run` entrypoint, which includes test discovery and
execution. Instead of calling this, we write a fixture to the test file
that calls the (new) `InterfaceTester.context` entrypoint, which handles
configuration, setup and teardown of the `interface_tester.Tester`
context. Currently, we install `pytest-interface-tester` from my fork,
but when [this
PR](https://github.com/canonical/pytest-interface-tester/pull/32) is
merged, we can switch to `pytest-interface-tester~=3.4`.

## Test failures

The tests for the `tracing` interface are currently failing. They're
also [failing in
charm-relation-interfaces](https://github.com/canonical/charm-relation-interfaces/actions/runs/18176547258/job/51743670143)
with exactly the same (`scenario`) error -- so it's not related to this
PR. This highlights the need to get interface tests running here
alongside the interface libraries, and in charm repos.

## Future work

A workflow to run interface tests from charm repos will be a separate
PR. This should perhaps be in its own repository so that it can be
versioned easily.