test(interfaces/tls_certificates): pull lib from PyPI for published version (#201)
This PR updates the `tls_certificates` interface tests to pull the
'published' version of the library from PyPI instead of embedding a copy
of the 4.22 Charmhub release. This wasn't possible in the PR that added
this library, as it had not yet been released to PyPI. Currently (and
previously), the code in the published and local versions are
essentially identical.