feat: add tls_certificates interface spec to charmlibs monorepo (#174)
This PR adds the `tls-certificates` interface [definition from
charm-relation-interfaces](https://github.com/canonical/charm-relation-interfaces/tree/main/interfaces/tls_certificates)
to this repo.

[**Preview docs
build**](https://canonical-ubuntu-documentation-library--174.com.readthedocs.build/charmlibs/reference/interfaces/tls-certificates/v1/)

At the same time, it renames the `interfaces/tls_certificates` directory
to `interfaces/tls-certificates` -- the interface in use seems to be
[tls-certificates](https://charmhub.io/integrations/tls-certificates).
We used the wrong directory name here initially because Charmhub hosts
the docs under
[tls_certificates](https://charmhub.io/integrations/tls_certificates).
The `CODEOWNERS` file has been updated to reflect this, which will
require review from `charmlibs-maintainers` in addition to review by the
`TLS` team.

---

`json` blocks in the `README.md` files have been made valid by replacing
invalid multi-line strings representing serialized `json` objects
(previously enclosed in `'[` and `]'`) with `"..."` and moving the
contents to a separate `json` block below. Maybe there's a better
approach here, open to changing it.

---

Some `pyright: ignore` statements are needed in the `schema.py` file,
because `pytest-interface-tester` lacks a `py.typed` marker, and because
schemas are expected to override fields of the Pydantic databag model it
defines in ways that the type checker dislikes.

Pyright dislikes how the model is used because it's not frozen (we can
probably fix this in a future release), and because it doesn't support
an idiom that Pydantic does -- see these closed Pyright issues:
- https://github.com/microsoft/pyright/issues/8766
- https://github.com/microsoft/pyright/issues/10347

Hopefully we can figure out a nicer approach as we move more interfaces
to this repo.