ci: add charm-interface-tests.yaml workflow (#209)
This PR adds a reusable workflow for running the interface tests for a
specific charm from an external repo. When this workflow is called, a
`charm` must be specified by name. The workflow then runs the subset of
the interface tests defined in `canonical/charmlibs@main` that apply to
that charm. The charm may also specify `charm_repo` and `charm_branch`
inputs to override those in the interfaces' charm config to run the
interface tests against (e.g.) PR branches.

## Implementation

The basic flow is similar to `interface-tests.yaml` -- in an `init` job,
we collect the interface test targets, calling `.scripts/ls.py` and
`.scripts/get-interface-test-targets.py`. Then in a `tests` matrix job
we run `.scripts/run-interface-tests.py`.

This workflow checks out `canonical/charmlibs@main` (hardcoded --
temporarily set to my fork for testing). It does this both to access the
workflow utility scripts, and to discover the interface definitions and
their tests.

`.scripts/get-interface-test-targets.py` has been updated to accept an
`--only-charm` argument to only output targets that include the charm
under test, and an `--exclude-charm` flag for shorter matrix job names.

`.scripts/run-interface-tests.py` has been updated to accept
`--charm-repo` and `--charm-branch` arguments which override the
interface's charm config.