docs: make template prompt accurate for interface libraries (#222)
When running `just init` or `just interface init`, the user is prompted
for their project name. The prompt previously said it should be the
package name, but this is only the case for `just init` -- with `just
interface init` it's important that the user instead specifies the
canonical interface name.

This PR updates the prompt displayed to account for both cases, using a
short generic prompt inside `cookiecutter`, but printing extra details
beforehand via the `just` recipe.