ci: actually use requested python with uv, and use system for functional (#223)
This PR updates the `justfile` to actually pass the `python` variable to
`uv` (!) -- we've been implicitly only using the system Python for some
time now.

Additionally, it updates the `functional` tests in CI to run with the
system Python -- previously we were using `matrix.python`, but at some
point `python` was removed from the matrix, so this was just setting
`python` to an empty string (which wasn't being passed to `uv`, so it
didn't fail). Since the `functional` tests are run on a matrix of Ubuntu
bases, it makes sense to just use the `python` from there.