ci: use up-to-date ref for git diff to detect changes vs PR target (#205)
This PR updates the `init` job in `ci.yaml` to compute an up-to-date
reference to the PR's target branch instead of using
`github.event.pull_request.base.sha`. Using
`github.event.pull_request.base.sha` is problematic for two reasons.

1. It is only updated when the PR is initially created and on force
pushing (or clicking the update branch button too, it seems). [Known
Github issue since
2023](https://github.com/orgs/community/discussions/59677).
2. When it is updated, a naive `git diff` with this reference will treat
new changes on `main` the same as changes introduced in the PR branch.
For example, if we merged a PR updating the `justfile`, CI will behave
as though the `justfile` was changed in this PR.

We compute the target with `git merge-base` to get the effects of the
[three-dot diff used by
Github](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/about-comparing-branches-in-pull-requests#three-dot-git-diff-comparison)
-- this means that we'll only see changes introduced by the PR, not
including changes made to the PR target that haven't been merged into
the PR yet.