docs: add CHANGELOG.md for charmlibs.nginx_k8s (#229)
This PR adds a `CHANGELOG.md` to the `charmlibs.nginx_k8s` library, with
an entry for the library's initial release.

It also updates the package's metadata with a link to this changelog,
and bumps the library version to `0.0.1.post0` to trigger a release.

Since we're updating the links, I've also made the documentation link
more specific.

## Context

Since we're not using Github releases in this repository, there isn't a
natural place to automatically publish release notes. Instead, each
library will maintain a checked-in `CHANGELOG.md` which serves this
purpose for users trying to figure out which version of the library they
might need.

When making future releases by changing the library version, CI will
complain if the `CHANGELOG.md` file isn't also modified in the same PR.