feat: add user facing interface tests workflow and docs (#235)
This PR adds the `charm-interface-tests.yaml` workflow from `main` to a
separate `interface-tests-v0` branch, changing the name to
`interface-tests.yaml` so that users can enjoy a shorter name.

The `interface-tests-v0` branch splits off from the initial commit to
this repo, so it doesn't have any of the other packages, interfaces,
docs, or CI. It just has:
- `.docs/.readthedocs.yaml`
- Updated to just have minimal contents to pass the docs build, so we
don't get spurious red Xs in CI.
- It doesn't seem to be possible to disable PR builds for this branch
only.
- `.github/workflows/interface-tests.yaml`
    - Copied from `charm-interface-tests.yaml` on `main`.
- Updated to checkout `canonical/charmlibs@main` instead of my original
PR fork (forgot to do this before merging #209).
- `.github/workflows/ci.yaml`
- Different from `main`, just runs the user facing
`interface-tests.yaml` workflow with a hardcoded list of charms.
- I have some ideas for populating the charms list dynamically, but we
can keep it simple for now.
- `.github/dependabot.yaml`
    - Copied from `main`, to keep dependencies up-to-date.
- `.gitignore` (unmodified from initial commit)
- `LICENSE` (unmodified)
- `README.md`
    - Provides information about the workflow.

This is the approach we decided on over #232

---

The interface test failures are unrelated -- the charms under test had
an update to a shared library recently that somehow breaks their fixture
setup. Assuming they use pinned dependencies, this would have been
caught by their using this workflow in their PR CI.

---------

Co-authored-by: Ben Hoyt <benhoyt@gmail.com>