ci: add zizmor (#236)
This PR adds `zizmor` to this repo, to provide security code scanning
for our workflows. It will be enabled as a required security check for
PRs after merging this PR.

The PR fixes the easy issues too -- narrowing permissions, avoiding
cache poisoning, pinning some workflows, and manually ignoring a
dangerous workflow trigger warning since it should be safe in practice.

I haven't fixed the `charm-interface-tests.yaml` workflow errors, as
that whole file will shortly be removed in #238.

The big outstanding thing is the extensive use of template expansion.
That's too big to fix in this PR. I'm not sure if we can merge as-is, or
if we should defer enabling the security checks as required till this is
addressed in a separate PR, or if we need to carve out an ignore in
`zizmor.yml` for now.