ci: remove charm-interface-tests from main and rename workflows (#238)
Now that the public-facing interface testing workflow is defined in the
`interface-tests-v0` branch, this PR removes the
`charm-interface-tests.yaml` from `main`.

Now that the public-facing interface testing workflow is named
`interface-tests.yaml`, this PR renames our internal workflow to
`test-interface.yaml`, renaming `tests.yaml` to `test-package.yaml` for
symmetry -- we use `packages` and `interfaces` as the two first class
citizens of this repo in a number of places.