docs: add a how-to for library users (#242)
This PR adds a how-to guide for library users. It steals some
suspiciously how-to-like content from the Charmhub-hosted libraries
explanation. It also steals some suspiciously library user focused
content from the how-to guide for distributing libraries, specifically
as git dependencies. It adds some fresh content for `charmlibs`
specifically on top.

Context: I'm working on a PR for the `charmcraft` docs that points
readers to `charmlibs` for library related guidance, and this page felt
like a clear gap in our docs.

---------

Co-authored-by: Dave Wilding <tech@dpw.me>