docs: fix silent failure to load doctree (#248)
This PR fixes a bug in the local `package_docs` Sphinx extension,
introduced when we started normalizing the URLs to use the canonical
distribution package name -- the package argument also needs to be
normalized when comparing to doctree names, otherwise we don't correctly
save the doctree for the current package docs being built. This results
in a blank page for that package.


[Live](https://documentation.ubuntu.com/charmlibs/reference/charmlibs/nginx-k8s/)
(blank)
[Preview
build](https://canonical-ubuntu-documentation-library--248.com.readthedocs.build/charmlibs/reference/charmlibs/nginx-k8s/)
(fixed)