docs: use markdown labels for interface reference docs (#249)
When I added labels to the automatically included interface reference
docs, I mistakenly prefixed the included content with a reStructuredText
label instead of a Markdown label. The rST label was then rendered in
the page as literal text. This PR corrects this mistake, using a
Markdown label, and also adds a label to the interface's index page.