ci: improvements for new contributors (#251)
This PR addresses some of the feedback we've received about the tooling
for this repository being a bit overwhelming for new contributors.

It adds a new `check` recipe that runs `lint`ing and `unit` tests, and
builds the `docs` for a given package.

It also moves the current, auto-generated `just` help to `just help`,
and adds a new hand-crafted help to be displayed when just running
`just`. Screenshots:

`CONTRIBUTING.md` has also been updated with similar quick start
information for working on a new package.

Additionally, per discussion in person, this PR removes `codespell` from
our CI, which resolves #246.
Currently this leaves us without any automated spellchecking, but we'll
try `vale` in #81.