docs(tls-certificates): correct url in readme (#258)
This PR corrects the docs URL in the
`charmlibs.interfaces.tls_certificates` readme. The package version is
bumped to trigger a 'post' release so the correction will display on
PyPI.

The error is due to our switching to normalizing all the reference docs
URLs on the docs site.

Fixes: #257

@canonical/tls: Feel free to combine this fix with #255 and close this
PR if that's easier.