docs: fix bad links in interface docs (#270)
We don't normally pay too much attention to the linkcheck job, as it's
often a bit flaky due to Charmhub. This has resulted in a number of bad
links being introduced with the bulk migration of interface docs.

For the most part, these are local links to JSON schemas which we don't
host. These have all been replaced with links to the interface's
Pydantic schema instead.

Additionally, the `openfga` docs included a deep link to the library's
source code in the Github repository, the layout of which has now
changed. I've replaced this with a hopefully more stable link to the
library on Charmhub.