docs: use fully qualified lib names in CSVs (#271)
This PR updates the Charmhub-hosted library names used in
`general-libs.csv` and `interface-libs.csv`.

Previously the CSV just included the library name in the `name` column,
and prefixed it with the charm name when rendering the docs. This worked
fine, but is a minor pain point when trying to look up library
information by name in the CSV, since some libraries have the same
library name (they're only guaranteed to be unique per charm).

Now the name column has the format `charms.<charm name>.<lib name>`,
with `charms.` being dropped when rendering the docs (so the displayed
form is the same as before).

I included the `charms.` prefix to make it clear just from the name
which libraries are Charmhub-hosted. When looking at the CSV, it's clear
from the URL and `kind`, but it's nice for the library name to make this
clear as well, for use cases like `lib: charms.<charm name>.<lib name>`
in `interface.yaml`.