feat: generate .interfaces.json file for Charmhub (#272)
This PR updates `.scripts/ls.py` to output additional interface metadata
that we want to expose to Charmhub.

This includes adding support for new fields in `interface.yaml`: `lib`,
`summary`, and `description`. These fields have been added to the
`tracing` interface as an example and to exercise the new code.

The `just interfaces-json` recipe has been added to write out the
`.interfaces.json` file, and a job has been added to CI to check that it
is up-to-date on PRs.