ci: interface JSON rename and CODEOWNERS (#281)
This PR renames `.interfaces.json` to `interfaces/index.json` to keep
the root directory less cluttered, and sets `interfaces/index.json` to
have no `CODEOWNERS`.
https://github.com/canonical/canonical-repo-automation/pull/531 makes
`interfaces-json-up-to-date` a required check for PRs, so the content
will always reflect the `CODEOWNERS` approved changes to the interface
directories' contents. This means that we don't need an additional
review for the file itself.